requirements.txt
anthropic
cryptography
google-generativeai
numpy
openai
python-dotenv
flask
